requirements.txt
flask
black
isort
autoflake
flask_jwt_extended
Flask-Mail
Flask-RESTful
Flask-SQLAlchemy
mysqlclient
python-dateutil
marshmallow
requests